Classic Curse Words Kata ------------------
We want to find out the most common swear words used in literature between 1500-1900 AD. We figured that leveraging the books available in the Gutenberg Project would be a great resource, so we will provide a way to get the data. We are also going to pre-determine curse words to mind for, both as our inputs. As usual, you may leverage any technology you would like. Since we are trying to solve our own problem, we don't really have expected output, so this will be more open-ended.